# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BougieBams Community & Events API
 * OpenAPI spec version: 0.1.0
 */
/**
 * Raw Stripe webhook event payload
 */
export interface WebhookPayload { [key: string]: unknown }

export interface HealthStatus {
  status: string;
}

export interface Event {
  id: number;
  title: string;
  description: string;
  date: string;
  /** @nullable */
  endDate?: string | null;
  location: string;
  /** @nullable */
  address?: string | null;
  /** Price in dollars */
  price: number;
  capacity: number;
  /** @nullable */
  spotsRemaining?: number | null;
  /** @nullable */
  imageUrl: string | null;
  category: string;
  /**
     * Comma-separated tags
     * @nullable
     */
  tags?: string | null;
  featured: boolean;
  published: boolean;
  /** @nullable */
  stripePriceId?: string | null;
  createdAt: string;
}

export interface EventInput {
  /** @minLength 1 */
  title: string;
  description: string;
  date: string;
  /** @nullable */
  endDate?: string | null;
  location: string;
  /** @nullable */
  address?: string | null;
  /** @minimum 0 */
  price: number;
  /** @minimum 1 */
  capacity: number;
  /** @nullable */
  imageUrl?: string | null;
  category: string;
  /** @nullable */
  tags?: string | null;
  featured?: boolean;
  published?: boolean;
  /** @nullable */
  stripePriceId?: string | null;
}

export interface EventUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  date?: string;
  /** @nullable */
  endDate?: string | null;
  location?: string;
  /** @nullable */
  address?: string | null;
  /** @minimum 0 */
  price?: number;
  /** @minimum 1 */
  capacity?: number;
  /** @nullable */
  imageUrl?: string | null;
  category?: string;
  /** @nullable */
  tags?: string | null;
  featured?: boolean;
  published?: boolean;
  /** @nullable */
  stripePriceId?: string | null;
}

export type RegistrationStatus = typeof RegistrationStatus[keyof typeof RegistrationStatus];


export const RegistrationStatus = {
  pending: 'pending',
  paid: 'paid',
  cancelled: 'cancelled',
} as const;

export interface Registration {
  id: number;
  eventId: number;
  firstName: string;
  lastName: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  quantity?: number;
  totalAmount?: number;
  /** @nullable */
  stripeSessionId?: string | null;
  status: RegistrationStatus;
  createdAt: string;
}

export interface RegistrationConfirmation {
  registration: Registration;
  event: Event;
}

export interface CheckoutSessionInput {
  eventId: number;
  /** @minLength 1 */
  firstName: string;
  /** @minLength 1 */
  lastName: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @minimum 1 */
  quantity: number;
}

export interface CheckoutSessionResponse {
  url: string;
  sessionId: string;
}

export interface RegistrationStats {
  totalRegistrations: number;
  totalRevenue: number;
  upcomingEvents: number;
  totalEvents: number;
  recentRegistrations?: Registration[];
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export interface ErrorEnvelope {
  error: string;
}

export type ListEventsParams = {
/**
 * Filter by event category
 * @nullable
 */
category?: string | null;
/**
 * Filter to featured events only
 * @nullable
 */
featured?: boolean | null;
/**
 * Filter to upcoming events only
 * @nullable
 */
upcoming?: boolean | null;
/**
 * Full-text search filter (title, description, location, category)
 * @nullable
 */
search?: string | null;
};

export type GetRegistrationBySessionParams = {
sessionId: string;
};

export type HandleStripeWebhook200 = {
  received?: boolean;
};