# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BougieBams Community & Events API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List all events
      description: Returns a list of all upcoming and past events
      parameters:
        - name: category
          in: query
          schema:
            type: ["string", "null"]
          description: Filter by event category
        - name: featured
          in: query
          schema:
            type: ["boolean", "null"]
          description: Filter to featured events only
        - name: upcoming
          in: query
          schema:
            type: ["boolean", "null"]
          description: Filter to upcoming events only
        - name: search
          in: query
          schema:
            type: ["string", "null"]
          description: Full-text search filter (title, description, location, category)
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create a new event (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get a single event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Event not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update an event (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Event not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted successfully
        "404":
          description: Event not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /events/{id}/registrations:
    get:
      operationId: listEventRegistrations
      tags: [registrations]
      summary: List registrations for an event (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of registrations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Registration"
  /registrations/checkout:
    post:
      operationId: createCheckoutSession
      tags: [registrations]
      summary: Create a Stripe Checkout session for event registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutSessionInput"
      responses:
        "200":
          description: Stripe Checkout session URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSessionResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Event not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /registrations/confirmation:
    get:
      operationId: getRegistrationBySession
      tags: [registrations]
      summary: Get registration details by Stripe session ID
      parameters:
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Registration details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationConfirmation"
        "404":
          description: Registration not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  schemas:
...
    Event:
      type: object
      required: [id, title, description, date, location, price, capacity, imageUrl, category, featured, published, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        date:
          type: string
          format: date-time
        endDate:
          type: ["string", "null"]
          format: date-time
        location:
          type: string
        address:
          type: ["string", "null"]
        price:
          type: number
          description: Price in dollars
        capacity:
          type: integer
        spotsRemaining:
          type: ["integer", "null"]
        imageUrl:
          type: ["string", "null"]
        category:
          type: string
        tags:
          type: ["string", "null"]
          description: Comma-separated tags
        featured:
          type: boolean
        published:
          type: boolean
        stripePriceId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    EventInput:
      type: object
      required: [title, description, date, location, price, capacity, category]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        date:
          type: string
          format: date-time
        endDate:
          type: ["string", "null"]
          format: date-time
        location:
          type: string
        address:
          type: ["string", "null"]
        price:
          type: number
          minimum: 0
        capacity:
          type: integer
          minimum: 1
        imageUrl:
          type: ["string", "null"]
        category:
          type: string
        tags:
          type: ["string", "null"]
        featured:
          type: boolean
        published:
          type: boolean
        stripePriceId:
          type: ["string", "null"]
    EventUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        date:
          type: string
          format: date-time
        endDate:
          type: ["string", "null"]
          format: date-time
        location:
          type: string
        address:
          type: ["string", "null"]
        price:
          type: number
          minimum: 0
        capacity:
          type: integer
          minimum: 1
        imageUrl:
          type: ["string", "null"]
        category:
          type: string
        tags:
          type: ["string", "null"]
        featured:
          type: boolean
        published:
          type: boolean
        stripePriceId:
          type: ["string", "null"]
    Registration:
      type: object
      required: [id, eventId, firstName, lastName, email, status, createdAt]
      properties:
        id:
          type: integer
        eventId:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: ["string", "null"]
        quantity:
          type: integer
        totalAmount:
          type: number
        stripeSessionId:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, paid, cancelled]
        createdAt:
          type: string
          format: date-time
    RegistrationConfirmation:
      type: object
      required: [registration, event]
      properties:
        registration:
          $ref: "#/components/schemas/Registration"
        event:
          $ref: "#/components/schemas/Event"
    CheckoutSessionInput:
      type: object
      required: [eventId, firstName, lastName, email, quantity]
      properties:
        eventId:
          type: integer
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        phone:
          type: ["string", "null"]
        quantity:
          type: integer
          minimum: 1
          default: 1
    CheckoutSessionResponse:
      type: object
      required: [url, sessionId]
      properties:
        url:
          type: string
          format: uri
        sessionId:
          type: string
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string